Entangle Code Review
A lot of time has come since my last commit. I had a busy school year, so the development was frozen.
Now I am FINALLY proudly announcing that the development of the 1.0 release is coming to its end.
I admit there still may be some bugs. I admit that I cannot guarantee total stability and security.
But that's now a totally different code - that's so far from what it was last year that it even looks like a different program!
Actually, I don't have the complete changelog to publish it here; but here is a brief explanation of my changes:
--- New Features ---
- Compression (useful for making backups)
- Other ways of selecting files
- Full installation on Linux
- New Makefile
--- Fixes ---
- Finished all translations.
- LOADS and LOADS of stability fixes!
- Code quality (read: beauty)
... and MUCH, MUCH MORE.
The last thing I would like to say here is that the worst thing that happened to the project and caused this year-long delay - that SecBlock assert on Windows - WAS NOT. MY. FAULT. I have avoided it by switching to TDM-GCC-64 compiler on Windows.
Thanks to everyone who helped me! All of you are amazing people I am lucky to know! :)

Enjoy (^_^),
Ilya.